Runtime/RTFuzz: Started to extend the master command to provide a network interface for controlling the fuzzing process. The interface uses JSON requests and responses to control the fuzzing process, like creating a new fuzzing run and suspending, resuming and stopping it. The JSON request contains everything required (except for the binary itself) for easier control of multiple masters on different hosts via a central control mechanism (later integration into the validation kit). [build fix]
svn:sync-xref-src-repo-rev: r123066

+++ src/VBox/Runtime/common/fuzz/fuzzmastercmd.cpp
@@ -1,4 +1,4 @@
-/* $Id: fuzzmastercmd.cpp 72571 2018-06-15 20:36:17Z alexander.eichner@oracle.com $ */
+/* $Id: fuzzmastercmd.cpp 72572 2018-06-15 20:40:14Z alexander.eichner@oracle.com $ */
 /** @file
  * IPRT - Fuzzing framework API, master command.
  */
@@ -242,7 +242,7 @@ static int rtFuzzCmdMasterFuzzRunProcessCfgSizeDef(size_t *pcbVal, const char *p
  * @param   cbDef               Default value if the item wasn't found.
  * @param   pErrInfo            Where to store the error information on failure, optional.
  */
-static int rtFuzzCmdMasterFuzzRunProcessCfgU32Def(uint32_t *pu32Val, const char *pszCfgItem, RTJSONVAL hJsonCfg, size_t u32Def, PRTERRINFO pErrInfo)
+static int rtFuzzCmdMasterFuzzRunProcessCfgU32Def(uint32_t *pu32Val, const char *pszCfgItem, RTJSONVAL hJsonCfg, uint32_t u32Def, PRTERRINFO pErrInfo)
 {
     int64_t i64Val = 0;
     int rc = RTJsonValueQueryIntegerByName(hJsonCfg, pszCfgItem, &i64Val);
